codejail_service/urls.py
"""
codejail_service URL Configuration.

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/3.2/topics/http/urls/

Examples:

Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  url(r'^$', views.home, name='home')

Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  url(r'^$', Home.as_view(), name='home')

Including another URLconf
    1. Add an import:  from blog import urls as blog_urls
    2. Add a URL to urlpatterns:  url(r'^blog/', include(blog_urls))
"""

from django.urls import include, re_path

from codejail_service.apps.api import urls as api_urls
from codejail_service.apps.core import views as core_views

urlpatterns = [
    re_path(r'^api/', include(api_urls)),
    re_path(r'^health/$', core_views.health, name='health'),
]
