requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    make upgrade
#
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
astroid==3.3.8
    # via
    #   pylint
    #   pylint-celery
cachetools==5.5.1
    # via tox
cffi==1.17.1
    # via
    #   -r requirements/base.txt
    #   pynacl
chardet==5.2.0
    # via tox
click==8.1.8
    # via
    #   -r requirements/base.txt
    #   click-log
    #   code-annotations
    #   edx-django-utils
    #   edx-lint
click-log==0.4.0
    # via edx-lint
code-annotations==2.2.0
    # via
    #   -r requirements/test.in
    #   edx-lint
colorama==0.4.6
    # via tox
coverage[toml]==7.6.10
    # via
    #   -r requirements/test.in
    #   pytest-cov
dill==0.3.9
    # via pylint
distlib==0.3.9
    # via virtualenv
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   django-waffle
    #   edx-django-utils
    #   edx-i18n-tools
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-waffle==4.2.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
edx-django-utils==7.1.0
    # via -r requirements/base.txt
edx-i18n-tools==1.6.3
    # via -r requirements/base.txt
edx-lint==5.5.0
    # via -r requirements/test.in
filelock==3.17.0
    # via
    #   tox
    #   virtualenv
gunicorn==23.0.0
    # via -r requirements/base.txt
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via pylint
jinja2==3.1.5
    # via code-annotations
lxml[html-clean]==5.3.0
    # via
    #   -r requirements/base.txt
    #   edx-i18n-tools
    #   lxml-html-clean
lxml-html-clean==0.4.1
    # via
    #   -r requirements/base.txt
    #   lxml
markupsafe==3.0.2
    # via jinja2
mccabe==0.7.0
    # via pylint
newrelic==10.4.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
openedx-atlas==0.6.2
    # via -r requirements/base.txt
packaging==24.2
    # via
    #   -r requirements/base.txt
    #   gunicorn
    #   pyproject-api
    #   pytest
    #   tox
path==16.16.0
    # via
    #   -r requirements/base.txt
    #   edx-i18n-tools
pbr==6.1.0
    # via
    #   -r requirements/base.txt
    #   stevedore
platformdirs==4.3.6
    # via
    #   pylint
    #   tox
    #   virtualenv
pluggy==1.5.0
    # via
    #   pytest
    #   tox
polib==1.2.0
    # via
    #   -r requirements/base.txt
    #   edx-i18n-tools
psutil==6.1.1
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycparser==2.22
    # via
    #   -r requirements/base.txt
    #   cffi
pylint==3.3.3
    # via
    #   edx-lint
    #   pylint-celery
    #   pylint-django
    #   pylint-plugin-utils
pylint-celery==0.3
    # via edx-lint
pylint-django==2.6.1
    # via edx-lint
pylint-plugin-utils==0.8.2
    # via
    #   pylint-celery
    #   pylint-django
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pyproject-api==1.9.0
    # via tox
pytest==8.3.4
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==6.0.0
    # via -r requirements/test.in
pytest-django==4.9.0
    # via -r requirements/test.in
python-slugify==8.0.4
    # via code-annotations
pytz==2024.2
    # via -r requirements/base.txt
pyyaml==6.0.2
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-i18n-tools
six==1.17.0
    # via edx-lint
sqlparse==0.5.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.4.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
text-unidecode==1.3
    # via python-slugify
tomlkit==0.13.2
    # via pylint
tox==4.24.1
    # via -r requirements/test.in
virtualenv==20.29.1
    # via tox
